+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed:
 - Disappearing keyboard on mobile
+- Improved text editor layout
 
 ## [0.8.223] - 2023-01-04
 ### Changed: